Moved wizards to a separate file. That makes more sense than keeping all the views in the same file.
